bubble sort

+++ pactice.c
@@ -27,10 +27,30 @@
 
 
 
+// #include <stdio.h>
+// int main(){
+//     system("cls");
+// int far ,cel ;
+// int lower, upper, step;
+// lower=0;
+// upper =300;
+// step=20;
+// far=lower;
+//  printf("far\t cel\n");
+// while (far<=upper){
+//     cel = 5*(far-32)/9;
+//     // printf("%d \t%d\n",far, cel);
+//     printf("%3d %6d\n",far, cel);
+//     far += step;
+// }
+// }
+
+
+
 #include <stdio.h>
 int main(){
     system("cls");
-int far ,cel ;
+float far ,cel ;
 int lower, upper, step;
 lower=0;
 upper =300;
@@ -40,7 +60,7 @@ far=lower;
 while (far<=upper){
     cel = 5*(far-32)/9;
     // printf("%d \t%d\n",far, cel);
-    printf("%3d %6d\n",far, cel);
+    printf("%3.0f %10.2f\n",far, cel);
     far += step;
 }
 }
